Configured project due to MVC pattern.
