+++ iOS6-HW13-Alexander.Strelkov/Model/Settings.swift
@@ -1,0 +1,35 @@
+//
+//  Settings.swift
+//  iOS6-HW13-Alexander.Strelkov
+//
+//  Created by Alexandr Strelkov on 17.06.2022.
+//
+
+import Foundation
+import UIKit
+
+struct Section {
+    let options: [SettingsOptionsType]
+}
+
+enum SettingsOptionsType {
+    case staticCell(model: SettingsOptions)
+    case switchCell(model: SettingsSwitchOption)
+}
+
+struct SettingsSwitchOption {
+    let title: String
+    let icon: UIImage?
+    let iconColor: UIColor
+    let handler: (() -> Void)
+    let isToggled: Bool
+}
+
+struct SettingsOptions {
+    let title: String
+    let icon: UIImage?
+    let iconColor: UIColor
+    let detail: String?
+    let notification: UIImage?
+    let handler: (() -> Void)
+}
