+++ iOS6-HW13-Alexander.Strelkov/View/SettingsView.swift
@@ -1,0 +1,110 @@
+//
+//  SettingsView.swift
+//  iOS6-HW13-Alexander.Strelkov
+//
+//  Created by Alexandr Strelkov on 31.05.2022.
+//
+
+import UIKit
+
+class SettingsView: UIView, UITableViewDelegate, UITableViewDataSource{
+    
+    
+    func configureView(with models: [Section]) {
+        self.models = models
+        tableView.reloadData()
+    }
+    
+    //MARK: -TableView
+    
+    private var tableView: UITableView = {
+        let table = UITableView(frame: .zero, style: .insetGrouped)
+        table.register(SettingTableViewCell.self, forCellReuseIdentifier: SettingTableViewCell.identifier)
+        table.register(ToggleTableViewCell.self, forCellReuseIdentifier: ToggleTableViewCell.identifier)
+        table.translatesAutoresizingMaskIntoConstraints = false
+        return table
+    }()
+    
+    private var models = [Section]()
+    
+    //MARK: -Initials:
+    
+    init() {
+        super.init(frame: .zero)
+        commonInit()
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        commonInit()
+    }
+    
+    private func commonInit() {
+        backgroundColor = .white
+        setupHierarchy()
+        setupLayout()
+    }
+    
+    //MARK: -Methods
+    
+    private func setupHierarchy() {
+        addSubview(tableView)
+        tableView.delegate = self
+        tableView.dataSource = self
+    }
+    
+    func setupLayout() {
+        tableView.leadingAnchor.constraint(equalTo: leadingAnchor).isActive = true
+        tableView.trailingAnchor.constraint(equalTo: trailingAnchor).isActive = true
+        tableView.topAnchor.constraint(equalTo: topAnchor).isActive = true
+        tableView.heightAnchor.constraint(equalTo: heightAnchor).isActive = true
+    }
+    
+    //MARK: -Delegate and DataSource
+    
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return models.count
+    }
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return models[section].options.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let model = models[indexPath.section].options[indexPath.row]
+        
+        switch model.self {
+        case .staticCell(let model):
+            guard let cell = tableView.dequeueReusableCell(
+                withIdentifier: SettingTableViewCell.identifier,
+                for: indexPath) as? SettingTableViewCell else {
+                    return UITableViewCell()
+                }
+            cell.configureCustomCell(with: model)
+            return cell
+            
+        case .switchCell(let model):
+            guard let cell = tableView.dequeueReusableCell(
+                withIdentifier: ToggleTableViewCell.identifier,
+                for: indexPath) as? ToggleTableViewCell else {
+                    return UITableViewCell()
+                }
+            cell.configureCustomCell(with: model)
+            return cell
+        }
+    }
+    
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        tableView.deselectRow(at: indexPath, animated: true)
+        let type = models[indexPath.section].options[indexPath.row]
+        switch type.self {
+        case .staticCell(let model):
+            model.handler()
+            print("Нажата ячейка \(model.title)")
+        case .switchCell(let model):
+            model.handler()
+            print("Нажата ячейка \(model.title)")
+        }
+    }
+}
+
